--- lion_master/src/app/layout/rdf-modelling/modelling-container/modelling-components/dashboard/dashboard.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { DashboardService } from './dashboard.service';
-
-describe('Service: Dashboard', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [DashboardService]
-    });
-  });
-
-  it('should ...', inject([DashboardService], (service: DashboardService) => {
-    expect(service).toBeTruthy();
-  }));
-});
